generating mixing bijections + test case

+++ src/main/java/cz/muni/fi/xklinec/whiteboxAES/generator/NormalGF2MatrixHolder.java
@@ -4,31 +4,29 @@
  */
 package cz.muni.fi.xklinec.whiteboxAES.generator;
 
-import org.bouncycastle.pqc.math.linearalgebra.GF2Matrix;
-
 /**
  *
  * @author ph4r05
  */
 public class NormalGF2MatrixHolder {
-    private long rank;
-    private long detetrminant;
+    private int rank;
+    private int detetrminant;
     private GF2MatrixEx P;
     private GF2MatrixEx Q;
 
-    public long getRank() {
+    public int getRank() {
         return rank;
     }
 
-    public void setRank(long rank) {
+    public void setRank(int rank) {
         this.rank = rank;
     }
 
-    public long getDetetrminant() {
+    public int getDetetrminant() {
         return detetrminant;
     }
 
-    public void setDetetrminant(long detetrminant) {
+    public void setDetetrminant(int detetrminant) {
         this.detetrminant = detetrminant;
     }
